tests/kvm: create a 'kvm' package
This will allow us to move stuff into 'kvm.*' modules, avoiding
proliferation of 'kvm_*.py' files in the top directory.

The existing kvm.py module was moved to 'kvm.test', and just the test
class (named 'kvm' too) is exposed by the 'kvm' package on
kvm/__init__.py, to allow the Autotest core to locate it.

Signed-off-by: Eduardo Habkost <ehabkost@raisama.net>