kvm: build: register always-failing installer object in case the build failed
No test code should ever hit this if the depdencies are set correctly on
the .cfg file. But in case the KVM build failed and dependencies are not
correctly set, we don't want any test that depends on the KVM install
information to incorrectly succeed as the modules aren't really
available.

Signed-off-by: Eduardo Habkost <ehabkost@raisama.net>