kvm: installer: install modules to 'build' directory
This changes the "source dir" and "git" KVM installers.

Instead of loading the modules directly from the source directory,
install them on 'build' so we can load them on other tests.

This also fixes some unintended side-effects of the previous install
code. On some cases, the modules were being installed into the system
/lib/modules directory, overwriting the system modules, and we don't
want that.

This also moves the installed-module loading code to SourceDirInstaller,
so that it loads the installed modules if necessary, instead of the
system-provided modules.

Signed-off-by: Eduardo Habkost <ehabkost@raisama.net>